+++ CHANGELOG.md
@@ -15,11 +15,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Earth2Studio base schema for dataframe sources
 - Added Planetary Computer data source for ECMWF IFS analysis data
 - Added accumulated variable support to NCAR ERA5 data source
+- Added Planetary Computer data source for GOES cloud and moisture imagery
+- Added ability to have seperate data and model cache locations via env variables
 - Added base reflectivity to MRMS data source
 
 ### Changed
 
+- Changed available date for ARCO data source to be dynamically updated
 - Changed ISD data frame return to master schema
+- Updated CBottle infill to mixture of model checkpoints
 
 ### Deprecated
 
@@ -28,12 +32,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Bug in cbottle datasource resulting identical samples for multiple samples
+- Bug in StormCast loading out-of-date model package, introduced in `5518edecbabee371c824b34f0f2ec269a4d6094f`
 
 ### Security
 
 ### Dependencies
 
 - Added pyarrow explicitly to core dependencies
+- Updated CBottle repo to NVlabs location
+- Updated Makani dependency hash to later version with pyproject toml fix
+- Updated ACE2 dependency hash with setuptools fix
 
 ## [0.12.1] - 2026-01-29